refactor: add obj to config, move type
